split out requests into own function files
