+++ routes/api/auth/router.js
@@ -2,12 +2,12 @@
   ~/auth - router
 ***********************************************************/
 
-const bcrypt = require ('bcryptjs')
-
 const {
   Router,
-  models : { users },
 } = require ('./__needs')
+
+const registerUser = require ('./registerUser')
+const loginUser = require ('./loginUser')
 
 /**************************************/
 
@@ -16,86 +16,10 @@ const router = Router ()
 /// requests ///
 
 router.route ('/auth/register')
-.post ((ri, ro) => {
-
-  const { username, password } = ri.body
-
-  const data = {
-    username,
-    hash : bcrypt.hashSync (password, 10),
-  }
-
-  users.push (data)
-  .then ((something) => {
-
-    ro
-    .status (201)
-    .json (something)
-
-  })
-  .catch ((error) => {
-
-    console.log (error)
-
-    ro
-    .status (500)
-    .json ({
-      'error' : {
-        'message' : `internal server error when registering new user`,
-        'method' : ri.method,
-        'route' : ri.originalUrl,
-      }
-    })
-
-  })
-
-})
+.post (registerUser)
 
 router.route ('/auth/login')
-.post ((ri, ro) => {
-
-  const { username, password } = ri.body
-
-  users.getBy ('username', username)
-  .then ((user) => {
-
-    if (user && bcrypt.compareSync (password, user.hash)) {
-
-      ro
-      .status (201)
-      .json ({
-        'message' : `Welcome, ${user.username}`,
-      })
-
-    } else {
-
-      ro
-      .status (401)
-      .json ({
-        'error' : {
-          'message' : `invalid user credentials`,
-        }
-      })
-
-    }
-
-  })
-  .catch ((error) => {
-
-    console.log (error)
-
-    ro
-    .status (500)
-    .json ({
-      'error' : {
-        'message' : `internal server error when logging in as user`,
-        'method' : ri.method,
-        'route' : ri.originalUrl,
-      }
-    })
-
-  })
-})
+.post (loginUser)
 
 /**************************************/
